Extract seed data module
